Allow Agimus to give sides in the great avocado war. (#422)
* Allow Agimus to give sides in the great avocado war.

We might want to check on some things.  There are error messages when a role message doesn’t have a banner, and `!q_update_role_messages` doesn’t delete messages.  (Although that might be because there is more than one bot on the test server, and that’s causing other issues.)

* Lessen the language

* clear flag for q_update_role_messages, cap avocado role names

* Clean up and make the header optional

---------

Co-authored-by: Peter of the Norse <rahmcoff@gmail.org>
Co-authored-by: zmattingly <z@matting.ly>