Bump helm chart versions in PR
This commit adds some new jobs to the PR pipelines.  The new jobs get the current `version` and `appVersion` from the helm chart, bumps them using the `semver` tool, commits the changes, pushes them back to the PR, and writes a comment on the PR suggesting the author pull down the changes.