Make backup function async
We kept hitting an error where the backup function would timeout and
from reading a little bit, I think it's due to the function not being
marked as async.  This corrects that oversight.

This commit also corrects the indentation in that function so that we
shouldn't hit any more undefined variables.