Automation to automatically bump semver (#538)
* Bump helm chart versions in PR

This commit adds some new jobs to the PR pipelines.  The new jobs get the current `version` and `appVersion` from the helm chart, bumps them using the `semver` tool, commits the changes, pushes them back to the PR, and writes a comment on the PR suggesting the author pull down the changes.

* Debugging yq issue

* Define outputs for job

* Trying to use step instead of needs

* Try it with new_version too

* Get new version by stdout

* Change from stdout to output per use-semver-tool code

* Make the change in the chart file and verify it

* Now get it working for appVersion

* Add in commiting and commenting

* Trying to get the proper branch checked out for the commit

* Add write permissions

* Bump helm chart version to 2.10.1

* Fixing the ping of a user on success

* Bump helm chart version to 2.10.2

* Bump helm version bump into own file

This allows us to only run the pipeline for bumping the helm version on open or reopen.  I'm not sure if that's what we want to do or if we want to run this on every PR change.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Mathew Fleisch <mathew.fleisch@gmail.com>
Co-authored-by: Zane Mattingly <z@matting.ly>
Co-authored-by: jp00p <asimovsghost@gmail.com>